+++ src/algorithm/nb_iter/osp/osp_shft.cpp
@@ -194,8 +194,8 @@ int OSHOP_shift_Ngbh::prepare_move( int typ )
 
 int OSHOP_shift_Ngbh::do_move()
  {
-   short predJ1;
-   short predM1;
+   int predJ1;
+   int predM1;
    *P[1]=*P[0];
 
    if (tabu_param[0][2]==1)
